+++ core/app/[locale]/(default)/product/[slug]/static/page.tsx
@@ -6,7 +6,7 @@ import { getChannelIdFromLocale } from '~/channels.config';
 import { client } from '~/client';
 import { graphql } from '~/client/graphql';
 import { revalidate as revalidateTarget } from '~/client/revalidate-target';
-import { locales } from '~/i18n';
+import { locales } from '~/i18n/routing';
 
 import ProductPage from '../page';
