+++ core/app/[locale]/maintenance/page.tsx
@@ -6,7 +6,7 @@ import { client } from '~/client';
 import { graphql } from '~/client/graphql';
 import { StoreLogo } from '~/components/store-logo';
 import { StoreLogoFragment } from '~/components/store-logo/fragment';
-import { LocaleType } from '~/i18n';
+import { LocaleType } from '~/i18n/routing';
 
 const MaintenancePageQuery = graphql(
   `
